Added UI, and created GameState class which control the waves and our health, coins e.t.c

+++ Assets/ScriptableObjects/Code/TowerSO.cs
@@ -9,5 +9,5 @@ public class TowerSO : ScriptableObject
     public float range;
     public float damage;
 
-    public float buildPrice;
+    public int buildPrice;
 }
